FIR/UAST: commonize when expression conversion
