+++ uast/uast-kotlin-fir/testData/legacyRenderLog/WhenIs.render.fir.txt
@@ -1,5 +1,15 @@
 public final class WhenIsKt {
     public static final fun foo(@null bar: java.lang.Object) : void {
-        return [!] UnknownKotlinExpression (WHEN)
+        return switch (bar)  {
+            it is java.lang.String -> {
+                yield bar
+            }
+
+            it !is java.lang.String -> {
+                yield "<error>"
+            }
+
+        }
+
     }
 }
