+++ uast/uast-kotlin-fir/testData/legacyRenderLog/WhenStringLiteral.log.fir.txt
@@ -5,7 +5,24 @@ UFile (package = )
             [!] UnknownKotlinExpression (CALL_EXPRESSION)
         UField (name = b)
             UAnnotation (fqName = org.jetbrains.annotations.NotNull)
-            [!] UnknownKotlinExpression (WHEN)
+            USwitchExpression
+                USimpleNameReferenceExpression (identifier = a)
+                UExpressionList (when)
+                    USwitchClauseExpressionWithBody
+                        ULiteralExpression (value = "abc")
+                        UExpressionList (when_entry)
+                            UYieldExpression
+                                ULiteralExpression (value = 1)
+                    USwitchClauseExpressionWithBody
+                        ULiteralExpression (value = "def")
+                        ULiteralExpression (value = "ghi")
+                        UExpressionList (when_entry)
+                            UYieldExpression
+                                ULiteralExpression (value = 2)
+                    USwitchClauseExpressionWithBody
+                        UExpressionList (when_entry)
+                            UYieldExpression
+                                ULiteralExpression (value = 3)
         UMethod (name = getA)
         UMethod (name = getB)
         UMethod (name = <no name provided>)
