+++ uast/uast-kotlin-fir/testData/legacyValues/WhenIs.values.fir.txt
@@ -4,5 +4,21 @@ UFile (package = ) [public final class WhenIsKt {...]
             UParameter (name = bar) [@null var bar: java.lang.Object]
                 UAnnotation (fqName = null) [@null]
             UBlockExpression [{...}] = Nothing
-                UReturnExpression [return [!] UnknownKotlinExpression (WHEN)] = Nothing
-                    [!] UnknownKotlinExpression (WHEN) [[!] UnknownKotlinExpression (WHEN)] = Undetermined
+                UReturnExpression [return switch (bar)  {...] = Nothing
+                    USwitchExpression [switch (bar)  {...] = Undetermined
+                        USimpleNameReferenceExpression (identifier = bar) [bar] = Undetermined
+                        UExpressionList (when) [    it is java.lang.String -> {...    ] = Undetermined
+                            USwitchClauseExpressionWithBody [it is java.lang.String -> {...] = Undetermined
+                                UBinaryExpressionWithType [it is java.lang.String] = Undetermined
+                                    USimpleNameReferenceExpression (identifier = it) [it] = Nothing
+                                    UTypeReferenceExpression (name = java.lang.String) [java.lang.String] = Undetermined
+                                UExpressionList (when_entry) [{...] = Undetermined
+                                    UYieldExpression [yield bar] = UYieldResult(Undetermined)
+                                        USimpleNameReferenceExpression (identifier = bar) [bar] = Undetermined
+                            USwitchClauseExpressionWithBody [it !is java.lang.String -> {...] = Undetermined
+                                UBinaryExpressionWithType [it !is java.lang.String] = Undetermined
+                                    USimpleNameReferenceExpression (identifier = it) [it] = Nothing
+                                    UTypeReferenceExpression (name = java.lang.String) [java.lang.String] = Undetermined
+                                UExpressionList (when_entry) [{...] = Undetermined
+                                    UYieldExpression [yield "<error>"] = UYieldResult("<error>")
+                                        ULiteralExpression (value = "<error>") ["<error>"] = "<error>"
