+++ uast/uast-kotlin-fir/testData/legacyValues/WhenStringLiteral.values.fir.txt
@@ -3,9 +3,26 @@ UFile (package = ) [public final class WhenStringLiteralKt {...]
         UField (name = a) [@org.jetbrains.annotations.Nullable private static final var a: java.lang.String = [!] UnknownKotlinExpression (CALL_EXPRESSION)]
             UAnnotation (fqName = org.jetbrains.annotations.Nullable) [@org.jetbrains.annotations.Nullable]
             [!] UnknownKotlinExpression (CALL_EXPRESSION) [[!] UnknownKotlinExpression (CALL_EXPRESSION)] = Undetermined
-        UField (name = b) [@org.jetbrains.annotations.NotNull private static final var b: int = [!] UnknownKotlinExpression (WHEN)]
+        UField (name = b) [@org.jetbrains.annotations.NotNull private static final var b: int = switch (a)  {...]
             UAnnotation (fqName = org.jetbrains.annotations.NotNull) [@org.jetbrains.annotations.NotNull]
-            [!] UnknownKotlinExpression (WHEN) [[!] UnknownKotlinExpression (WHEN)] = Undetermined
+            USwitchExpression [switch (a)  {...] = Phi((long)1, (long)2)
+                USimpleNameReferenceExpression (identifier = a) [a] = Undetermined
+                UExpressionList (when) [    "abc" -> {...    ] = Phi((long)1, (long)2)
+                    USwitchClauseExpressionWithBody ["abc" -> {...] = Undetermined
+                        ULiteralExpression (value = "abc") ["abc"] = "abc"
+                        UExpressionList (when_entry) [{...] = Undetermined
+                            UYieldExpression [yield 1] = UYieldResult((long)1)
+                                ULiteralExpression (value = 1) [1] = (long)1
+                    USwitchClauseExpressionWithBody ["def", "ghi" -> {...] = Undetermined
+                        ULiteralExpression (value = "def") ["def"] = "def"
+                        ULiteralExpression (value = "ghi") ["ghi"] = "ghi"
+                        UExpressionList (when_entry) [{...] = Undetermined
+                            UYieldExpression [yield 2] = UYieldResult((long)2)
+                                ULiteralExpression (value = 2) [2] = (long)2
+                    USwitchClauseExpressionWithBody [ -> {...] = Undetermined
+                        UExpressionList (when_entry) [{...] = Undetermined
+                            UYieldExpression [yield 3] = UYieldResult((long)3)
+                                ULiteralExpression (value = 3) [3] = (long)3
         UMethod (name = getA) [public static final fun getA() : java.lang.String = UastEmptyExpression]
         UMethod (name = getB) [public static final fun getB() : int = UastEmptyExpression]
         UMethod (name = <no name provided>) [public static final fun <no name provided>() : void {...}]
