+++ uast/uast-kotlin/src/org/jetbrains/uast/kotlin/KotlinConverter.kt
@@ -15,10 +15,7 @@ import org.jetbrains.kotlin.asJava.classes.KtLightClass
 import org.jetbrains.kotlin.asJava.elements.*
 import org.jetbrains.kotlin.asJava.findFacadeClass
 import org.jetbrains.kotlin.asJava.toLightClass
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocLink
 import org.jetbrains.kotlin.kdoc.psi.impl.KDocName
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocSection
-import org.jetbrains.kotlin.kdoc.psi.impl.KDocTag
 import org.jetbrains.kotlin.lexer.KtTokens
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
@@ -29,23 +26,6 @@ import org.jetbrains.uast.kotlin.expressions.*
 import org.jetbrains.uast.kotlin.psi.*
 
 internal object KotlinConverter : BaseKotlinConverter {
-    internal tailrec fun unwrapElements(element: PsiElement?): PsiElement? = when (element) {
-        is KtValueArgumentList -> unwrapElements(element.parent)
-        is KtValueArgument -> unwrapElements(element.parent)
-        is KtDeclarationModifierList -> unwrapElements(element.parent)
-        is KtContainerNode -> unwrapElements(element.parent)
-        is KtSimpleNameStringTemplateEntry -> unwrapElements(element.parent)
-        is KtLightParameterList -> unwrapElements(element.parent)
-        is KtTypeElement -> unwrapElements(element.parent)
-        is KtSuperTypeList -> unwrapElements(element.parent)
-        is KtFinallySection -> unwrapElements(element.parent)
-        is KtAnnotatedExpression -> unwrapElements(element.parent)
-        is KtWhenConditionWithExpression -> unwrapElements(element.parent)
-        is KDocLink -> unwrapElements(element.parent)
-        is KDocSection -> unwrapElements(element.parent)
-        is KDocTag -> unwrapElements(element.parent)
-        else -> element
-    }
 
     override fun convertAnnotation(annotationEntry: KtAnnotationEntry, givenParent: UElement?): UAnnotation {
         return KotlinUAnnotation(annotationEntry, givenParent)
@@ -286,48 +266,6 @@ internal object KotlinConverter : BaseKotlinConverter {
 
             else -> expr<UExpression>(build(::UnknownKotlinExpression))
         }}
-    }
-
-    internal fun convertWhenCondition(
-        condition: KtWhenCondition,
-        givenParent: UElement?,
-        requiredType: Array<out Class<out UElement>>
-    ): UExpression? {
-        val project = condition.project
-        val service = ServiceManager.getService(project, BaseKotlinUastResolveProviderService::class.java)
-
-        return with(requiredType) {
-            when (condition) {
-                is KtWhenConditionInRange -> expr<UBinaryExpression> {
-                    KotlinCustomUBinaryExpression(condition, givenParent).apply {
-                        leftOperand = KotlinStringUSimpleReferenceExpression("it", this, service)
-                        operator = when {
-                            condition.isNegated -> KotlinBinaryOperators.NOT_IN
-                            else -> KotlinBinaryOperators.IN
-                        }
-                        rightOperand = convertOrEmpty(condition.rangeExpression, this)
-                    }
-                }
-                is KtWhenConditionIsPattern -> expr<UBinaryExpression> {
-                    KotlinCustomUBinaryExpressionWithType(condition, givenParent).apply {
-                        operand = KotlinStringUSimpleReferenceExpression("it", this, service)
-                        operationKind = when {
-                            condition.isNegated -> KotlinBinaryExpressionWithTypeKinds.NEGATED_INSTANCE_CHECK
-                            else -> UastBinaryExpressionWithTypeKind.INSTANCE_CHECK
-                        }
-                        val typeRef = condition.typeReference
-                        typeReference = typeRef?.let {
-                            KotlinUTypeReferenceExpression(it, this, service) { typeRef.toPsiType(this, boxed = true) }
-                        }
-                    }
-                }
-
-                is KtWhenConditionWithExpression ->
-                    condition.expression?.let { convertExpression(it, givenParent, requiredType) }
-
-                else -> expr<UExpression> { UastEmptyExpression(givenParent) }
-            }
-        }
     }
 
     private fun convertEnumEntry(original: KtEnumEntry, givenParent: UElement?): UElement? {
